+++ traffic_ops/client/deliveryservice.go
@@ -73,13 +73,6 @@ const (
 	// See Also: https://traffic-control-cdn.readthedocs.io/en/latest/api/v2/deliveryservices_xmlid_xmlid_sslkeys.html
 	API_DELIVERY_SERVICE_XMLID_SSL_KEYS = API_DELIVERY_SERVICES + "/xmlid/%s/sslkeys"
 
-	// API_DELIVERY_SERVICES_HOSTNAME_SSL_KEYS is the API path on which Traffic Ops serves information about
-	// and functionality relating to the SSL keys used by a Delivery Service identified by its 'host name'.
-	// It is intended to be used with fmt.Sprintf to insert its required path parameter (namely the 'host
-	// name' of the Delivery Service of interest).
-	// See Also: https://traffic-control-cdn.readthedocs.io/en/latest/api/v2/deliveryservices_hostname_hostname_sslkeys.html
-	API_DELIVERY_SERVICE_HOSTNAME_SSL_KEYS = API_DELIVERY_SERVICES + "/hostname/%s/sslkeys"
-
 	// API_DELIVERY_SERVICE_URI_SIGNING_KEYS is the API path on which Traffic Ops serves information
 	// about and functionality relating to the URI-signing keys used by a Delivery Service identified
 	// by its XMLID. It is intended to be used with fmt.Sprintf to insert its required path parameter
@@ -98,11 +91,6 @@ const (
 	// 'regex' (Regular Expression) information.
 	// See Also: https://traffic-control-cdn.readthedocs.io/en/latest/api/v2/deliveryservices_regexes.html
 	API_DELIVERY_SERVICES_REGEXES = apiBase + "/deliveryservices_regexes"
-
-	// API_DELIVERY_SERVICE_MATCHES is the API path on which Traffic Ops serves information about
-	// the "matches" or "patterns" used by Delivery Services.
-	// See Also: https://traffic-control-cdn.readthedocs.io/en/latest/api/v2/deliveryservice_matches.html
-	API_DELIVERY_SERVICE_MATCHES = apiBase + "/deliveryservice_matches"
 
 	// API_SERVER_DELIVERY_SERVICES is the API path on which Traffic Ops serves functionality
 	// related to the associations a specific server and its assigned Delivery Services. It is
@@ -426,29 +414,6 @@ func (to *Session) GetDeliveryServiceSSLKeysByID(XMLID string) (*tc.DeliveryServ
 	}
 
 	return &data.Response, reqInf, nil
-}
-
-// GetDeliveryServiceSSLKeysByHostname returns information about the SSL keys used by the Delivery
-// Service identified by the passed 'host name'.
-func (to *Session) GetDeliveryServiceSSLKeysByHostname(hostname string) (*tc.DeliveryServiceSSLKeys, ReqInf, error) {
-	var data tc.DeliveryServiceSSLKeysResponse
-	reqInf, err := get(to, fmt.Sprintf(API_DELIVERY_SERVICE_HOSTNAME_SSL_KEYS, hostname), &data)
-	if err != nil {
-		return nil, reqInf, err
-	}
-
-	return &data.Response, reqInf, nil
-}
-
-// GetDeliveryServiceMatches returns the "matches" or "patterns" used by all (tenant-visible)
-// Delivery Services.
-func (to *Session) GetDeliveryServiceMatches() ([]tc.DeliveryServicePatterns, ReqInf, error) {
-	resp := tc.DeliveryServiceMatchesResponse{}
-	reqInf, err := get(to, API_DELIVERY_SERVICE_MATCHES, &resp)
-	if err != nil {
-		return nil, reqInf, err
-	}
-	return resp.Response, reqInf, nil
 }
 
 // GetDeliveryServicesEligible returns the servers eligible for assignment to the Delivery